2024 Day 3 [896/2366]
Lost a bunch of time on Part 2 because I was still reading the file line by line and put "doMuls" inside the loop, didn't realize that, then went ahead and hand edited the newlines out once I realized what was up.

Now the "doMuls" is in the wrong place but I've also added a "ReadEntireFile" function in case I need to just treat the whole file as one big string next time.