2024 Day 5 [2169/960] (Runs in <2ms)
Part 1 was simply checking that the rule for each number in the list contains none of the previous numbers in the list (the rule map is "for a given page, which pages come after") and summing the middle values.
Part 2 required sorting the bad ranges so they're good (using the rules again) and summing the middle values.

I lost a minute due to I accidentally put the answer to the sample problem into the answer box for part 1.