2024 Day 6 Initial solution [474/259] (Runs in 3.2s)
Part 1 was fine, simply run through the grid until you either hit a wall (and turn) or go out of bounds, marking in a set every position
Part 2 is slow, it basically tries to put a block at every (non-starting) visited position and runs the guard until we reach a square going the same direction as previously found.

There are faster ways to do it.