2024 Day 7 [late to the day] (runs in 39.5ms)
I agonized a bunch over how to memoize this thing before just trying the brute force and it turned out it was fine.
Then I did the same thing for part 2 (since 3^12 >> 2^12) but it also was fine. So I burned a bunch of my own time for nothing.

Took me 39 minutes to finish part 1 after opening the puzzle, and 44 total. Still not great, I should have just tried to blast through the first part.