2024 Day 7 Even More optimization (2.3ms now)
Realized that if you did the operations from right to left (Basically undoing the operations as you go) you can early out faster (especially on multiply/concatenation) because you can check the modulo of the operation and immediately fail if  it's not a proper multiple.

That's probably good now. Way faster than the initial time.