2024 Day 13 [3178/1043] (runs in 0.8ms)
I should have had this much faster tonight, I saw the linear equation, said "oh nice that's useful" and then flubbed the math and decided to do something else. When that didn't finish in a reasonable time I went back to the linear equations, got it right this time, and it meant that P1 -> P2 was less than a minute difference.