2024 Day 16 Revamped again
STILL not worth the time it took but now it runs in 1.44ms instead of ... 3.5.

Now among other things there's no memory allocation on the fly  (unless there's a surprising number of paths, I suppose)

A second grid is used to keep track of the paths that are in use, and because the walk is going from lowest -> highest cost we actually will never go both directions along a single path, so each end points to the same path (in a vector, as an index). Paths get committed to their direction when the shortest walk along them pops out of the queue (so a path may be queued in both directions, but it'll commit as soon as we get a shortest path).